Allow configuring build functionality with build tags
