+++ pkg/backends/tcp.go
@@ -1,3 +1,6 @@
+// +build !no_tcp_backend
+// +build !no_backends
+
 package backends
 
 import (
