+++ pkg/backends/udp.go
@@ -1,3 +1,6 @@
+// +build !no_udp_backend
+// +build !no_backends
+
 package backends
 
 import (
