+++ pkg/controlsvc/controlsvc.go
@@ -1,3 +1,5 @@
+// +build !no_controlsvc
+
 package controlsvc
 
 import (
@@ -16,25 +18,6 @@ import (
 	"strings"
 	"sync"
 )
-
-// ControlCommandType is a type of command that can be run from the control service
-type ControlCommandType interface {
-	InitFromString(string) (ControlCommand, error)
-	InitFromJSON(map[string]interface{}) (ControlCommand, error)
-}
-
-// ControlCommand is an instance of a command that is being run from the control service
-type ControlCommand interface {
-	ControlFunc(*netceptor.Netceptor, ControlFuncOperations) (map[string]interface{}, error)
-}
-
-// ControlFuncOperations provides callbacks for control services to take actions
-type ControlFuncOperations interface {
-	BridgeConn(message string, bc io.ReadWriteCloser, bcName string) error
-	ReadFromConn(message string, out io.Writer) error
-	WriteToConn(message string, in chan []byte) error
-	Close() error
-}
 
 // sockControl implements the ControlFuncOperations interface that is passed back to control functions
 type sockControl struct {
