+++ pkg/services/unix_proxy.go
@@ -1,3 +1,6 @@
+// +build !no_proxies
+// +build !no_services
+
 package services
 
 import (
