+++ pkg/utils/flock_windows.go
@@ -1,4 +1,4 @@
-//+build windows
+// +build windows
 
 package utils
