+++ pkg/utils/unixsock_windows.go
@@ -1,4 +1,4 @@
-//+build windows
+// +build windows
 
 package utils
