+++ pkg/workceptor/command_detach_unixlike.go
@@ -1,4 +1,4 @@
-//+build !windows
+// +build !windows,!no_workceptor
 
 package workceptor
