+++ pkg/workceptor/command_detach_windows.go
@@ -1,4 +1,4 @@
-//+build windows
+// +build windows,!no_workceptor
 
 package workceptor
