+++ pkg/workceptor/json_test.go
@@ -1,3 +1,5 @@
+// +build !no_workceptor
+
 package workceptor
 
 import (
