+++ pkg/workceptor/lock_test.go
@@ -1,3 +1,5 @@
+// +build !no_workceptor
+
 package workceptor
 
 import (
