+++ pkg/workceptor/stdio_utils.go
@@ -1,3 +1,5 @@
+// +build !no_workceptor
+
 package workceptor
 
 import (
