+++ pkg/workceptor/workunitbase.go
@@ -1,3 +1,5 @@
+// +build !no_workceptor
+
 package workceptor
 
 import (
@@ -48,45 +50,12 @@ func WorkStateToString(workState int) string {
 	}
 }
 
-// WorkUnit represents a local unit of work
-type WorkUnit interface {
-	Init(w *Workceptor, unitID string, workType string, params string)
-	ID() string
-	UnitDir() string
-	StatusFileName() string
-	StdoutFileName() string
-	Save() error
-	Load() error
-	UpdateBasicStatus(state int, detail string, stdoutSize int64)
-	UpdateFullStatus(statusFunc func(*StatusFileData))
-	LastUpdateError() error
-	Status() *StatusFileData
-	Start() error
-	Restart() error
-	Cancel() error
-	Release(force bool) error
-}
-
 // ErrPending is returned when an operation hasn't succeeded or failed yet
 var ErrPending = fmt.Errorf("operation pending")
 
 // IsPending returns true if the error is an ErrPending
 func IsPending(err error) bool {
 	return err == ErrPending
-}
-
-// NewWorkerFunc represents a factory of WorkUnit instances
-type NewWorkerFunc func() WorkUnit
-
-// StatusFileData is the structure of the JSON data saved to a status file.
-// This struct should only contain value types, except for ExtraData.
-type StatusFileData struct {
-	State      int
-	Detail     string
-	StdoutSize int64
-	WorkType   string
-	Params     string
-	ExtraData  interface{}
 }
 
 // BaseWorkUnit includes data common to all work units, and partially implements the WorkUnit interface
